backend/pyproject.toml
[project]
name = "sports-spots-backend"
version = "0.1.0"
description = "Minimal Flask backend for Sports Spots"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.3.0",
]

[tool.uv]
dev-dependencies = []
